Move all error-related classes into error package
