+++ app/src/androidTest/java/org/schabi/newpipe/error/ErrorInfoTest.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.report;
+package org.schabi.newpipe.error;
 
 import android.os.Parcel;
