+++ app/src/main/AndroidManifest.xml
@@ -85,7 +85,7 @@
             android:name=".ExitActivity"
             android:label="@string/general_error"
             android:theme="@android:style/Theme.NoDisplay" />
-        <activity android:name=".report.ErrorActivity" />
+        <activity android:name=".error.ErrorActivity" />
 
         <!-- giga get related -->
         <activity
@@ -106,7 +106,7 @@
         </activity>
 
         <activity
-            android:name=".ReCaptchaActivity"
+            android:name=".error.ReCaptchaActivity"
             android:label="@string/recaptcha" />
 
         <provider
