+++ app/src/main/java/org/schabi/newpipe/error/AcraReportSenderFactory.java
@@ -1,18 +1,21 @@
-package org.schabi.newpipe.report;
+package org.schabi.newpipe.error;
 
 import android.content.Context;
 
 import androidx.annotation.NonNull;
 
-import org.acra.data.CrashReportData;
+import com.google.auto.service.AutoService;
+
+import org.acra.config.CoreConfiguration;
 import org.acra.sender.ReportSender;
-import org.schabi.newpipe.R;
+import org.acra.sender.ReportSenderFactory;
+import org.schabi.newpipe.App;
 
 /*
- * Created by Christian Schabesberger  on 13.09.16.
+ * Created by Christian Schabesberger on 13.09.16.
  *
  * Copyright (C) Christian Schabesberger 2015 <chris.schabesberger@mailbox.org>
- * AcraReportSender.java is part of NewPipe.
+ * AcraReportSenderFactory.java is part of NewPipe.
  *
  * NewPipe is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -28,12 +31,14 @@ import org.schabi.newpipe.R;
  * along with NewPipe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-public class AcraReportSender implements ReportSender {
-
-    @Override
-    public void send(@NonNull final Context context, @NonNull final CrashReportData report) {
-        ErrorActivity.reportError(context, report,
-                ErrorInfo.make(UserAction.UI_ERROR, "none",
-                        "App crash, UI failure", R.string.app_ui_crash));
+/**
+ * Used by ACRA in {@link App}.initAcra() as the factory for report senders.
+ */
+@AutoService(ReportSenderFactory.class)
+public class AcraReportSenderFactory implements ReportSenderFactory {
+    @NonNull
+    public ReportSender create(@NonNull final Context context,
+                               @NonNull final CoreConfiguration config) {
+        return new AcraReportSender();
     }
 }
