+++ app/src/main/java/org/schabi/newpipe/settings/SettingMigrations.java
@@ -7,9 +7,9 @@ import android.util.Log;
 import androidx.preference.PreferenceManager;
 
 import org.schabi.newpipe.R;
-import org.schabi.newpipe.report.ErrorActivity;
-import org.schabi.newpipe.report.ErrorInfo;
-import org.schabi.newpipe.report.UserAction;
+import org.schabi.newpipe.error.ErrorActivity;
+import org.schabi.newpipe.error.ErrorInfo;
+import org.schabi.newpipe.error.UserAction;
 
 import static org.schabi.newpipe.MainActivity.DEBUG;
