+++ app/src/main/res/layout/activity_error.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".report.ErrorActivity">
+    tools:context=".error.ErrorActivity">
 
     <include
         layout="@layout/toolbar_layout"
