+++ app/src/test/java/org/schabi/newpipe/error/ErrorActivityTest.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.report;
+package org.schabi.newpipe.error;
 
 import android.app.Activity;
